@@ -5,10 +5,11 @@
 name: c_json_example
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.13.0 <3.0.0'
 
 dependencies:
   ffi: ^1.0.0
+  path: ^1.8.0
 
 dev_dependencies:
   ffigen:
@@ -24,6 +25,8 @@ ffigen:
     include-directives:
       - '**cJSON.h'
   comments: false
+  typedef-map:
+    'size_t': 'IntPtr'
 
   preamble: |
     // Copyright (c) 2009-2017 Dave Gamble and cJSON contributors
