@@ -5,7 +5,7 @@
 name: libclang_example
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.13.0 <3.0.0'
 
 dev_dependencies:
   ffigen:
@@ -61,6 +61,11 @@ ffigen:
     long long: 8
     unsigned long long: 8
     enum: 4
+
+  # These can be used to map any typedef to a specific native type.
+  typedef-map:
+    'size_t': 'IntPtr'
+    'time_t': 'Int64'
 
   # Default is 'NativeLibrary'
   name: 'LibClang'
