@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'binding_string.dart';
-import 'typedef.dart';
 import 'writer.dart';
 
 /// Base class for all Bindings.
@@ -28,8 +27,8 @@ abstract class Binding {
     this.dartDoc,
   });
 
-  /// Return typedef dependencies.
-  List<Typedef> getTypedefDependencies(Writer w) => const [];
+  /// Get all dependencies, including itself and save them in [dependencies].
+  void addDependencies(Set<Binding> dependencies);
 
   /// Converts a Binding to its actual string representation.
   ///
