@@ -5,7 +5,7 @@
 import 'package:ffigen/src/code_generator.dart';
 
 import 'compound.dart';
-import 'typedef.dart';
+import 'typealias.dart';
 import 'writer.dart';
 
 class _SubType {
@@ -39,10 +39,18 @@ enum BroadType {
   Compound,
   NativeFunction,
 
+  /// Represents a function type.
+  FunctionType,
+
+  /// Represents a typealias.
+  Typealias,
+
   /// Represents a Dart_Handle.
   Handle,
 
-  /// Stores its element type in NativeType as only those are supported.
+  Enum,
+
+  /// Represents an Array type.
   ConstantArray,
   IncompleteArray,
 
@@ -68,11 +76,23 @@ class Type {
     SupportedNativeType.IntPtr: _SubType(c: 'IntPtr', dart: 'int'),
   };
 
+  /// Enum type is mapped to [SupportedNativeType.Int32].
+  static const enumNativeType = SupportedNativeType.Int32;
+
   /// Reference to the [Compound] binding this type refers to.
   Compound? compound;
 
-  /// Reference to the [Typedef] this type refers to.
-  Typedef? nativeFunc;
+  /// Reference to the [NativeFunc] this type refers to.
+  NativeFunc? nativeFunc;
+
+  /// Reference to the [Typealias] this type refers to.
+  Typealias? typealias;
+
+  /// Reference to the [FunctionType] this type refers to.
+  FunctionType? functionType;
+
+  /// Reference to the [EnumClass] this type refers to.
+  EnumClass? enumClass;
 
   /// For providing [SupportedNativeType] only.
   final SupportedNativeType? nativeType;
@@ -93,8 +113,11 @@ class Type {
     required this.broadType,
     this.child,
     this.compound,
+    this.enumClass,
     this.nativeType,
     this.nativeFunc,
+    this.typealias,
+    this.functionType,
     this.length,
     this.unimplementedReason,
   });
@@ -111,8 +134,18 @@ class Type {
   factory Type.union(Union union) {
     return Type._(broadType: BroadType.Compound, compound: union);
   }
-  factory Type.nativeFunc(Typedef nativeFunc) {
+  factory Type.enumClass(EnumClass enumClass) {
+    return Type._(broadType: BroadType.Enum, enumClass: enumClass);
+  }
+  factory Type.functionType(FunctionType functionType) {
+    return Type._(
+        broadType: BroadType.FunctionType, functionType: functionType);
+  }
+  factory Type.nativeFunc(NativeFunc nativeFunc) {
     return Type._(broadType: BroadType.NativeFunction, nativeFunc: nativeFunc);
+  }
+  factory Type.typealias(Typealias typealias) {
+    return Type._(broadType: BroadType.Typealias, typealias: typealias);
   }
   factory Type.nativeType(SupportedNativeType nativeType) {
     return Type._(broadType: BroadType.NativeType, nativeType: nativeType);
@@ -143,6 +176,26 @@ class Type {
     return Type._(broadType: BroadType.Handle);
   }
 
+  /// Get all dependencies of this type and save them in [dependencies].
+  void addDependencies(Set<Binding> dependencies) {
+    switch (broadType) {
+      case BroadType.Compound:
+        return compound!.addDependencies(dependencies);
+      case BroadType.NativeFunction:
+        return nativeFunc!.addDependencies(dependencies);
+      case BroadType.FunctionType:
+        return functionType!.addDependencies(dependencies);
+      case BroadType.Typealias:
+        return typealias!.addDependencies(dependencies);
+      case BroadType.Enum:
+        return enumClass!.addDependencies(dependencies);
+      default:
+        if (child != null) {
+          return child!.addDependencies(dependencies);
+        }
+    }
+  }
+
   /// Get base type for any type.
   ///
   /// E.g int** has base [Type] of int.
@@ -167,16 +220,33 @@ class Type {
     }
   }
 
-  bool get isPrimitive =>
-      (broadType == BroadType.NativeType || broadType == BroadType.Boolean);
+  /// Get base typealias type.
+  ///
+  /// Returns itself if it's not a Typealias.
+  Type getBaseTypealiasType() {
+    if (broadType == BroadType.Typealias) {
+      return typealias!.type.getBaseTypealiasType();
+    } else {
+      return this;
+    }
+  }
+
+  /// Function to check if the dart and C type string are same.
+  bool sameDartAndCType(Writer w) => getCType(w) == getDartType(w);
 
   /// Returns true if the type is a [Compound] and is incomplete.
-  bool get isIncompleteCompound =>
-      (broadType == BroadType.Compound &&
-          compound != null &&
-          compound!.isInComplete) ||
-      (broadType == BroadType.ConstantArray &&
-          getBaseArrayType().isIncompleteCompound);
+  bool get isIncompleteCompound {
+    final baseTypealiasType = getBaseTypealiasType();
+    if (baseTypealiasType == this) {
+      return (broadType == BroadType.Compound &&
+              compound != null &&
+              compound!.isInComplete) ||
+          (broadType == BroadType.ConstantArray &&
+              getBaseArrayType().isIncompleteCompound);
+    } else {
+      return baseTypealiasType.isIncompleteCompound;
+    }
+  }
 
   String getCType(Writer w) {
     switch (broadType) {
@@ -186,8 +256,10 @@ class Type {
         return '${w.ffiLibraryPrefix}.Pointer<${child!.getCType(w)}>';
       case BroadType.Compound:
         return '${compound!.name}';
+      case BroadType.Enum:
+        return '${w.ffiLibraryPrefix}.${_primitives[enumNativeType]!.c}';
       case BroadType.NativeFunction:
-        return '${w.ffiLibraryPrefix}.NativeFunction<${nativeFunc!.name}>';
+        return '${w.ffiLibraryPrefix}.NativeFunction<${nativeFunc!.type.getCType(w)}>';
       case BroadType
           .IncompleteArray: // Array parameters are treated as Pointers in C.
         return '${w.ffiLibraryPrefix}.Pointer<${child!.getCType(w)}>';
@@ -198,8 +270,12 @@ class Type {
         return '${w.ffiLibraryPrefix}.${_primitives[SupportedNativeType.Uint8]!.c}';
       case BroadType.Handle:
         return '${w.ffiLibraryPrefix}.Handle';
-      default:
-        throw Exception('cType unknown');
+      case BroadType.FunctionType:
+        return functionType!.getCType(w);
+      case BroadType.Typealias:
+        return typealias!.name;
+      case BroadType.Unimplemented:
+        throw UnimplementedError('C type unknown for ${broadType.toString()}');
     }
   }
 
@@ -211,8 +287,10 @@ class Type {
         return '${w.ffiLibraryPrefix}.Pointer<${child!.getCType(w)}>';
       case BroadType.Compound:
         return '${compound!.name}';
+      case BroadType.Enum:
+        return _primitives[enumNativeType]!.dart;
       case BroadType.NativeFunction:
-        return '${w.ffiLibraryPrefix}.NativeFunction<${nativeFunc!.name}>';
+        return '${w.ffiLibraryPrefix}.NativeFunction<${nativeFunc!.type.getDartType(w)}>';
       case BroadType
           .IncompleteArray: // Array parameters are treated as Pointers in C.
         return '${w.ffiLibraryPrefix}.Pointer<${child!.getCType(w)}>';
@@ -223,13 +301,87 @@ class Type {
         return _primitives[SupportedNativeType.Uint8]!.dart;
       case BroadType.Handle:
         return 'Object';
-      default:
-        throw Exception('dart type unknown for ${broadType.toString()}');
+      case BroadType.FunctionType:
+        return functionType!.getDartType(w);
+      case BroadType.Typealias:
+        // Typealias cannot be used by name in Dart types unless both the C and
+        // Dart type of the underlying types are same.
+        if (typealias!.type.sameDartAndCType(w)) {
+          return typealias!.name;
+        } else {
+          return typealias!.type.getDartType(w);
+        }
+      case BroadType.Unimplemented:
+        throw UnimplementedError(
+            'dart type unknown for ${broadType.toString()}');
     }
   }
 
   @override
   String toString() {
     return 'Type: $broadType';
+  }
+}
+
+/// Represents a function type.
+class FunctionType {
+  final Type returnType;
+  final List<Parameter> parameters;
+
+  FunctionType({
+    required this.returnType,
+    required this.parameters,
+  });
+
+  String getCType(Writer w, {bool writeArgumentNames = true}) {
+    final sb = StringBuffer();
+
+    // Write return Type.
+    sb.write(returnType.getCType(w));
+
+    // Write Function.
+    sb.write(' Function(');
+    sb.write(parameters.map<String>((p) {
+      return '${p.type.getCType(w)} ${writeArgumentNames ? p.name : ""}';
+    }).join(', '));
+    sb.write(')');
+
+    return sb.toString();
+  }
+
+  String getDartType(Writer w, {bool writeArgumentNames = true}) {
+    final sb = StringBuffer();
+
+    // Write return Type.
+    sb.write(returnType.getDartType(w));
+
+    // Write Function.
+    sb.write(' Function(');
+    sb.write(parameters.map<String>((p) {
+      return '${p.type.getDartType(w)} ${writeArgumentNames ? p.name : ""}';
+    }).join(', '));
+    sb.write(')');
+
+    return sb.toString();
+  }
+
+  void addDependencies(Set<Binding> dependencies) {
+    returnType.addDependencies(dependencies);
+    parameters.forEach((p) => p.type.addDependencies(dependencies));
+  }
+}
+
+/// Represents a NativeFunction<Function>.
+class NativeFunc {
+  final Type type;
+
+  NativeFunc.fromFunctionType(FunctionType functionType)
+      : type = Type.functionType(functionType);
+
+  NativeFunc.fromFunctionTypealias(Typealias typealias)
+      : type = Type.typealias(typealias);
+
+  void addDependencies(Set<Binding> dependencies) {
+    type.addDependencies(dependencies);
   }
 }
