@@ -2,16 +2,14 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:ffi';
-
 import 'package:ffigen/src/code_generator.dart';
+import 'package:ffigen/src/header_parser/includer.dart';
+import 'package:ffigen/src/header_parser/type_extractor/extractor.dart';
 import 'package:logging/logging.dart';
 
 import '../clang_bindings/clang_bindings.dart' as clang_types;
 import '../data.dart';
-import '../sub_parsers/enumdecl_parser.dart';
 import '../utils.dart';
-import 'compounddecl_parser.dart';
 
 final _logger = Logger('ffigen.header_parser.typedefdecl_parser');
 
@@ -32,88 +30,75 @@ final _logger = Logger('ffigen.header_parser.typedefdecl_parser');
 ///
 /// typedef A D; // Typeref.
 /// ```
-class _ParsedTypedef {
-  Binding? binding;
+class _ParsedTypealias {
+  Typealias? typealias;
   String? typedefName;
   bool typedefToPointer = false;
-  _ParsedTypedef();
+  _ParsedTypealias();
 }
 
-final _stack = Stack<_ParsedTypedef>();
+final _stack = Stack<_ParsedTypealias>();
 
 /// Parses a typedef declaration.
-Binding? parseTypedefDeclaration(clang_types.CXCursor cursor) {
-  _stack.push(_ParsedTypedef());
-
-  /// Check if typedef declaration is to a pointer.
-  _stack.top.typedefToPointer =
-      (clang.clang_getTypedefDeclUnderlyingType(cursor).kind ==
-          clang_types.CXTypeKind.CXType_Pointer);
-
-  // Name of typedef.
-  _stack.top.typedefName = cursor.spelling();
-  final resultCode = clang.clang_visitChildren(
-    cursor,
-    Pointer.fromFunction(_typedefdeclarationCursorVisitor,
-        clang_types.CXChildVisitResult.CXChildVisit_Break),
-    nullptr,
-  );
-
-  visitChildrenResultChecker(resultCode);
-  return _stack.pop().binding;
-}
-
-/// Visitor for extracting binding for a TypedefDeclarations of a
-/// [clang.CXCursorKind.CXCursor_TypedefDecl].
 ///
-/// Visitor invoked on cursor of type declaration returned by
-/// [clang.clang_getTypeDeclaration_wrap].
-int _typedefdeclarationCursorVisitor(clang_types.CXCursor cursor,
-    clang_types.CXCursor parent, Pointer<Void> clientData) {
-  try {
-    _logger.finest(
-        'typedefdeclarationCursorVisitor: ${cursor.completeStringRepr()}');
+/// Returns `null` if the typedef could not be generated or has been excluded
+/// by the config.
+Typealias? parseTypedefDeclaration(
+  clang_types.CXCursor cursor, {
+  bool pointerReference = false,
+}) {
+  _stack.push(_ParsedTypealias());
+  final typedefName = cursor.spelling();
+  final typedefUsr = cursor.usr();
+  if (shouldIncludeTypealias(typedefUsr, typedefName)) {
+    final ct = clang.clang_getTypedefDeclUnderlyingType(cursor);
+    final s = getCodeGenType(ct, pointerReference: pointerReference);
 
-    switch (clang.clang_getCursorKind(cursor)) {
-      case clang_types.CXCursorKind.CXCursor_StructDecl:
-        if (_stack.top.typedefToPointer &&
-            bindingsIndex.isSeenStruct(cursor.usr())) {
-          // Skip a typedef pointer if struct is seen.
-          _stack.top.binding = bindingsIndex.getSeenStruct(cursor.usr());
-        } else {
-          // This will update the name of struct if already seen.
-          _stack.top.binding = parseCompoundDeclaration(
-            cursor,
-            CompoundType.struct,
-            name: _stack.top.typedefName,
-          );
-        }
-        break;
-      case clang_types.CXCursorKind.CXCursor_UnionDecl:
-        if (_stack.top.typedefToPointer &&
-            bindingsIndex.isSeenUnion(cursor.usr())) {
-          // Skip a typedef pointer if struct is seen.
-          _stack.top.binding = bindingsIndex.getSeenUnion(cursor.usr());
-        } else {
-          // This will update the name of struct if already seen.
-          _stack.top.binding = parseCompoundDeclaration(
-            cursor,
-            CompoundType.union,
-            name: _stack.top.typedefName,
-          );
-        }
-        break;
-      case clang_types.CXCursorKind.CXCursor_EnumDecl:
-        _stack.top.binding =
-            parseEnumDeclaration(cursor, name: _stack.top.typedefName);
-        break;
-      default:
-        _logger.finest('typedefdeclarationCursorVisitor: Ignored');
+    if (bindingsIndex.isSeenUnsupportedTypealias(typedefUsr)) {
+      // Do not process unsupported typealiases again.
+    } else if (s.broadType == BroadType.Unimplemented) {
+      _logger
+          .fine("Skipped Typedef '$typedefName': Unimplemented type referred.");
+      bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    } else if (s.broadType == BroadType.Compound &&
+        s.compound!.originalName == typedefName) {
+      // Ignore typedef if it refers to a compound with the same original name.
+      bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+      _logger.fine(
+          "Skipped Typedef '$typedefName': Name matches with referred struct/union.");
+    } else if (s.broadType == BroadType.Enum) {
+      // Ignore typedefs to Enum.
+      bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+      _logger.fine("Skipped Typedef '$typedefName': typedef to enum.");
+    } else if (s.broadType == BroadType.Handle) {
+      // Ignore typedefs to Handle.
+      _logger.fine("Skipped Typedef '$typedefName': typedef to Dart Handle.");
+      bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    } else if (s.broadType == BroadType.ConstantArray ||
+        s.broadType == BroadType.IncompleteArray) {
+      // Ignore typedefs to Constant Array.
+      _logger.fine("Skipped Typedef '$typedefName': typedef to array.");
+      bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    } else if (s.broadType == BroadType.Boolean) {
+      // Ignore typedefs to Boolean.
+      _logger.fine("Skipped Typedef '$typedefName': typedef to bool.");
+      bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    } else {
+      // Create typealias.
+      _stack.top.typealias = Typealias(
+        usr: typedefUsr,
+        originalName: typedefName,
+        name: config.typedefs.renameUsingConfig(typedefName),
+        type: s,
+        dartDoc: getCursorDocComment(cursor),
+      );
+      bindingsIndex.addTypealiasToSeen(typedefUsr, _stack.top.typealias!);
     }
-  } catch (e, s) {
-    _logger.severe(e);
-    _logger.severe(s);
-    rethrow;
   }
-  return clang_types.CXChildVisitResult.CXChildVisit_Continue;
+
+  if (bindingsIndex.isSeenTypealias(typedefUsr)) {
+    _stack.top.typealias = bindingsIndex.getSeenTypealias(typedefUsr);
+  }
+
+  return _stack.pop().typealias;
 }
