@@ -28,21 +28,14 @@ class Bindings {
         0;
   }
 
-  late final _test1_ptr = _lookup<ffi.NativeFunction<_c_test1>>('test1');
-  late final _dart_test1 _test1 = _test1_ptr.asFunction<_dart_test1>();
+  late final _test1_ptr = _lookup<
+      ffi.NativeFunction<
+          ffi.Uint8 Function(ffi.Uint8, ffi.Pointer<ffi.Uint8>)>>('test1');
+  late final _test1 =
+      _test1_ptr.asFunction<int Function(int, ffi.Pointer<ffi.Uint8>)>();
 }
 
 class test2 extends ffi.Struct {
   @ffi.Uint8()
   external int a;
 }
-
-typedef _c_test1 = ffi.Uint8 Function(
-  ffi.Uint8 a,
-  ffi.Pointer<ffi.Uint8> b,
-);
-
-typedef _dart_test1 = int Function(
-  int a,
-  ffi.Pointer<ffi.Uint8> b,
-);
