@@ -21,25 +21,17 @@ class Bindings {
     return _a();
   }
 
-  late final _a_ptr = _lookup<ffi.NativeFunction<_c_a>>('a');
-  late final _dart_a _a = _a_ptr.asFunction<_dart_a>();
+  late final _a_ptr = _lookup<ffi.NativeFunction<ffi.Void Function()>>('a');
+  late final _a = _a_ptr.asFunction<void Function()>();
 
   void b() {
     return _b();
   }
 
-  late final _b_ptr = _lookup<ffi.NativeFunction<_c_b>>('b');
-  late final _dart_b _b = _b_ptr.asFunction<_dart_b>();
+  late final _b_ptr = _lookup<ffi.NativeFunction<ffi.Void Function()>>('b');
+  late final _b = _b_ptr.asFunction<void Function()>();
 }
 
 class c extends ffi.Opaque {}
 
 class d extends ffi.Opaque {}
-
-typedef _c_a = ffi.Void Function();
-
-typedef _dart_a = void Function();
-
-typedef _c_b = ffi.Void Function();
-
-typedef _dart_b = void Function();
