@@ -27,6 +27,8 @@ ${strings.headers}:
   ${strings.includeDirectives}:
     - '**cJSON.h'
 ${strings.comments}: false
+${strings.typedefmap}:
+  'size_t': 'IntPtr'
 ${strings.preamble}: |
   // Copyright (c) 2009-2017 Dave Gamble and cJSON contributors
   //
@@ -52,7 +54,7 @@ ${strings.preamble}: |
 
       matchLibraryWithExpected(
         library,
-        ['test', 'debug_generated', 'c_json.dart'],
+        ['test', 'debug_generated', 'example_c_json.dart'],
         ['example', 'c_json', config.output],
       );
     });
