@@ -36,12 +36,11 @@ ${strings.unions}:
         ''') as yaml.YamlMap),
       );
     });
-
     test('Expected bindings', () {
       matchLibraryWithExpected(actual, [
         'test',
         'debug_generated',
-        'opaque_dependencies_test_output.dart'
+        'header_parser_opaque_dependencies_test_output.dart'
       ], [
         'test',
         'header_parser_tests',
